+++ Adj/Algebra/Category.hs
@@ -1,0 +1,11 @@
+module Adj.Algebra.Category where
+
+import Adj.Algebra.Semigroupoid (Semigroupoid)
+
+{- |
+> * Left identity: identity . f ≡ f
+> * Right identity: f . identity ≡ f
+-}
+
+class Semigroupoid morphism => Category morphism where
+	identity :: morphism a a