+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * Define `Adj.Algebra` module
 * Define `Adj.Program` module
 * Define `Adj.Algebra.Semigroupoid` module
+* Define `Adj.Algebra.Category` module